requirements.txt
mcp>=1.2.0
anthropic>=0.8.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.1.0
pytesseract>=0.3.13
# 牛只检测依赖（可选）
opencv-python>=4.8.0
supervision>=0.16.0
ultralytics>=8.0.0